requirements.txt
albumentations==1.4.4
argcomplete==3.3.0
async-timeout==4.0.3
blinker==1.7.0
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cv==1.0.0
cycler==0.12.1
Cython==3.0.10
easydict==1.13
faiss-cpu==1.8.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy
onnx==1.16.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
Pillow==10.3.0
pipx==1.5.0
prettytable==3.10.0
protobuf==5.26.1
Pygments==2.17.2
pyodbc==5.1.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
redis==5.0.4
requests==2.31.0
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.4.0
tifffile==2024.4.24
tomli==2.0.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.2
wcwidth==0.2.13
Werkzeug==3.0.2
